Add placeholder array
Add an array that will be filled with placeholder characters. This will be displayed to the user. As the user guesses letters correctly, the letters will be placed in the correct spots in the placeholder array. The print statement has been modified to help test if the correct number of placeholders are displayed regardless of the length of the guess word.